Update to build against mm.0.6.0
mm's Audio library has changed to no longer take an offset and length as
arguments to buffer operations. When writing a large number of samples,
we therefore need to create two buffers: one of length 1024 frames for
writing the bulk of the data, and one of length `(samples mod 1024)` for
writing the last few frames.

Update the travis build matrix to only build OCaml versions 4.08 and
later, since mm.0.6.0 requires at least 4.08.